Fix stdout flushing in musl
In this commit, we replace our old patch by a more robust patch
series. We temporary remove the call to `ioctl` in `__stdout_write.c`
which causes an assertion error due to invalid argument order. In
addition, `__stdio_write` is fixed by updating a condition to avoid an
infinite loop since the last caracter of the buffer is skipped when
the `cnt` variable is tested with the `iov_len` variable.

Signed-off-by: Gaulthier Gain <gaulthier.gain@uliege.be>